requirements.txt
feedparser>=6.0
trafilatura>=1.6
lxml_html_clean>=0.1
anthropic>=0.39
resend>=2.0
python-dateutil>=2.8
croniter>=1.3
